backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
python-dotenv==1.0.1
pyproj>=3.6.0
numpy>=1.26.0
